requirements.txt
pandas
numpy
flask
scikit-learn
mysql.connector
sendgrid
